Move client predicates into their own package
