+++ src/main/java/seedu/address/model/client/predicates/CategoryContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.client;
+package seedu.address.model.client.predicates;
 
 import java.util.List;
 import java.util.Set;
@@ -6,6 +6,7 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.client.Client;
 
 /**
  * Tests that a {@code Client}'s {@code Category} matches any of the keywords given.
