+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -9,9 +9,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.model.client.Client;
-import seedu.address.model.client.ClientSatisfyAllPredicate;
-import seedu.address.model.client.NameContainsKeywordsPredicate;
+import seedu.address.model.client.predicates.ClientSatisfyAllPredicate;
 
 public class FindCommandParserTest {
