+++ src/test/java/seedu/address/model/client/ProductPreferenceContainsKeywordsPredicateTest.java
@@ -9,6 +9,7 @@ import java.util.Collections;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
+import seedu.address.model.client.predicates.ProductPreferenceContainsKeywordsPredicate;
 import seedu.address.testutil.PersonBuilder;
 
 public class ProductPreferenceContainsKeywordsPredicateTest {
